+++ src/main/kotlin/com/github/zero9178/mlirods/language/highlighting/TableGenSemanticTokensAnnotator.kt
@@ -5,9 +5,10 @@ import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.Annotator
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.openapi.editor.DefaultLanguageHighlighterColors
+import com.intellij.openapi.project.DumbAware
 import com.intellij.psi.PsiElement
 
-private class TableGenSemanticTokensAnnotator : Annotator {
+private class TableGenSemanticTokensAnnotator : Annotator, DumbAware {
     override fun annotate(element: PsiElement, holder: AnnotationHolder) {
         when (element) {
             is TableGenLetDirective -> {