+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 # MLIR ODS Changelog
 
 ## [Unreleased]
+
+## [0.3.0] - 2025-02-04
 ### Added
 - Syntax highlighting of fields in `let`-expressions, record fields and field accesses
 - Color page in `Editor -> Color Scheme -> TableGen` 
 - Added editor folding of code literals, braced bodies and angle brackets   
+- Syntax highlighting of bang operators
 
 ## [0.2.0] - 2025-01-16
 ### Added